# perf: Better optimize `ListSourcegraphDotComIndexableRepos`