--- a/internal/database/repos.go
+++ b/internal/database/repos.go
@@ -1190,15 +1190,18 @@ func (s *repoStore) ListSourcegraphDotComIndexableRepos(ctx context.Context, opt
 		tr.Finish()
 	}()
 
-	var where, joins []*sqlf.Query
-
+	var joins, where []*sqlf.Query
 	if opts.CloneStatus != types.CloneStatusUnknown {
-		joins = append(joins, sqlf.Sprintf(
-			"JOIN gitserver_repos gr ON gr.repo_id = repo.id",
-		))
-		where = append(where, sqlf.Sprintf(
-			"gr.clone_status = %s", opts.CloneStatus,
-		))
+		if opts.CloneStatus == types.CloneStatusCloned {
+			// Optimization: if we're asking for all cloned repos, then we can more quickly filter out
+			// repos that have a different status. This takes into account repo <-> gitserver_repos rows
+			// are 1-to-1, and the number of cloned repos are MUCH larger on the DotCom instance where
+			// this query is running.
+			joins = append(joins, sqlf.Sprintf("LEFT JOIN gitserver_repos gr ON gr.repo_id = repo.id AND gr.clone_status <> %s", types.CloneStatusCloned))
+			where = append(where, sqlf.Sprintf("gr.repo_id IS NULL"))
+		} else {
+			joins = append(joins, sqlf.Sprintf("JOIN gitserver_repos gr ON gr.repo_id = repo.id AND gr.clone_status = %s", opts.CloneStatus))
+		}
 	}
 
 	if len(where) == 0 {
@@ -1232,22 +1235,24 @@ func (s *repoStore) ListSourcegraphDotComIndexableRepos(ctx context.Context, opt
 	return results, nil
 }
 
+// N.B. This query's exact conditions are mirrored in the Postgres index
+// repo_dotcom_indexable_repos_idx. Any substantial changes to this query
+// may require an associated index redefinition.
 const listSourcegraphDotComIndexableReposQuery = `
 SELECT
-	repo.id, repo.name, repo.stars
+	repo.id,
+	repo.name,
+	repo.stars
 FROM repo
 %s
 WHERE
+	deleted_at IS NULL AND
+	blocked IS NULL AND
 	(
 		(repo.stars >= %s AND NOT COALESCE(fork, false) AND NOT archived)
 		OR
 		lower(repo.name) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)'
-	)
-	AND
-	deleted_at IS NULL
-	AND
-	blocked IS NULL
-	AND
+	) AND
 	%s
 ORDER BY stars DESC NULLS LAST
 `