--- a/migrations/frontend/1667950421_speed_up_listsourcegraphdotcomindexablerepos/down.sql
+++ b/migrations/frontend/1667950421_speed_up_listsourcegraphdotcomindexablerepos/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS repo_dotcom_indexable_repos_idx;