--- a/migrations/frontend/1667950421_speed_up_listsourcegraphdotcomindexablerepos/metadata.yaml
+++ b/migrations/frontend/1667950421_speed_up_listsourcegraphdotcomindexablerepos/metadata.yaml
@@ -1,0 +1,3 @@
+name: Speed up ListSourcegraphDotComIndexableRepos
+parents: [1667917030]
+createIndexConcurrently: true