--- a/migrations/frontend/1667952974_speed_up_listsourcegraphdotcomindexablerepos_part_2/down.sql
+++ b/migrations/frontend/1667952974_speed_up_listsourcegraphdotcomindexablerepos_part_2/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS gitserver_repos_not_explicitly_cloned_idx;