--- a/migrations/frontend/1667952974_speed_up_listsourcegraphdotcomindexablerepos_part_2/metadata.yaml
+++ b/migrations/frontend/1667952974_speed_up_listsourcegraphdotcomindexablerepos_part_2/metadata.yaml
@@ -1,0 +1,3 @@
+name: Speed up ListSourcegraphDotComIndexableRepos (part 2)
+parents: [1667950421]
+createIndexConcurrently: true