--- a/migrations/frontend/1667952974_speed_up_listsourcegraphdotcomindexablerepos_part_2/up.sql
+++ b/migrations/frontend/1667952974_speed_up_listsourcegraphdotcomindexablerepos_part_2/up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY IF NOT EXISTS gitserver_repos_not_explicitly_cloned_idx ON gitserver_repos (repo_id) WHERE clone_status <> 'cloned';