**Before**:

The following new queries (`not_cloned`/`cloned`, respectively) have the following query plans:

```sql
SELECT
    repo.id,
    repo.name,
    repo.stars
FROM repo
JOIN gitserver_repos gr ON gr.repo_id = repo.id
WHERE
   deleted_at IS NULL AND blocked IS NULL AND (
        (repo.stars >= 5 AND NOT COALESCE(fork, false) AND NOT archived)
        OR
        (lower(repo.name) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)')
    )
    AND gr.clone_status =  'not_cloned'
ORDER BY
    stars DESC NULLS LAST;
```

```
                                                                                                                                       QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Merge  (cost=1006499.41..1049436.90 rows=368010 width=46) (actual time=4059.684..4163.265 rows=9093 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Sort  (cost=1005499.38..1005959.40 rows=184005 width=46) (actual time=4050.370..4052.963 rows=3031 loops=3)
         Sort Key: repo.stars DESC NULLS LAST
         Sort Method: quicksort  Memory: 390kB
         Worker 0:  Sort Method: quicksort  Memory: 358kB
         Worker 1:  Sort Method: quicksort  Memory: 344kB
         ->  Parallel Hash Join  (cost=799165.95..989408.71 rows=184005 width=46) (actual time=3139.409..4051.087 rows=3031 loops=3)
               Hash Cond: (gr.repo_id = repo.id)
               ->  Parallel Index Only Scan using gitserver_repos_not_cloned_status_idx on gitserver_repos gr  (cost=0.43..187584.45 rows=1012855 width=4) (actual time=0.048..537.718 rows=793012 loops=3)
                     Heap Fetches: 1033064
               ->  Parallel Hash  (cost=792112.25..792112.25 rows=564262 width=46) (actual time=3133.421..3133.424 rows=937633 loops=3)
                     Buckets: 4194304 (originally 2097152)  Batches: 1 (originally 1)  Memory Usage: 268704kB
                     ->  Parallel Bitmap Heap Scan on repo  (cost=26246.27..792112.25 rows=564262 width=46) (actual time=1061.553..2581.419 rows=937633 loops=3)
                           Recheck Cond: (((stars >= 5) AND (deleted_at IS NULL) AND (blocked IS NULL) AND (NOT archived) AND (NOT COALESCE(fork, false))) OR ((lower((name)::text) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)'::text) AND (deleted_at IS NULL) AND (blocked IS NULL)))
                           Rows Removed by Index Recheck: 2392
                           Heap Blocks: exact=211806
                           ->  BitmapOr  (cost=26246.27..26246.27 rows=1379405 width=0) (actual time=793.253..793.255 rows=0 loops=1)
                                 ->  Bitmap Index Scan on repo_stars_indexable  (cost=0.00..24377.83 rows=1246250 width=0) (actual time=516.163..516.163 rows=2661778 loops=1)
                                       Index Cond: (stars >= 5)
                                 ->  Bitmap Index Scan on erik_test_remove7  (cost=0.00..1191.33 rows=133155 width=0) (actual time=277.088..277.089 rows=170719 loops=1)
                                       Index Cond: (lower((name)::text) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)'::text)
 Planning Time: 2.912 ms
 Execution Time: 4163.891 ms
(25 rows)
```

```sql
SELECT
    repo.id,
    repo.name,
    repo.stars
FROM repo
JOIN gitserver_repos gr ON gr.repo_id = repo.id
WHERE
   deleted_at IS NULL AND blocked IS NULL AND (
        (repo.stars >= 5 AND NOT COALESCE(fork, false) AND NOT archived)
        OR
        (lower(repo.name) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)')
    )
    AND gr.clone_status =  'cloned'
ORDER BY
    stars DESC NULLS LAST;
```

```
                                                                                                                                       QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Merge  (cost=1127457.89..1217820.67 rows=774484 width=46) (actual time=6746.873..8423.563 rows=2803728 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Sort  (cost=1126457.87..1127425.97 rows=387242 width=46) (actual time=6731.426..6918.842 rows=934576 loops=3)
         Sort Key: repo.stars DESC NULLS LAST
         Sort Method: quicksort  Memory: 112985kB
         Worker 0:  Sort Method: quicksort  Memory: 114564kB
         Worker 1:  Sort Method: quicksort  Memory: 113826kB
         ->  Parallel Hash Join  (cost=799165.95..1090516.24 rows=387242 width=46) (actual time=3192.357..6412.817 rows=934576 loops=3)
               Hash Cond: (gr.repo_id = repo.id)
               ->  Parallel Index Only Scan using gitserver_repos_cloned_status_idx on gitserver_repos gr  (cost=0.43..285755.34 rows=2131576 width=4) (actual time=0.036..1921.238 rows=1691811 loops=3)
                     Heap Fetches: 3106406
               ->  Parallel Hash  (cost=792112.25..792112.25 rows=564262 width=46) (actual time=3186.648..3186.652 rows=937633 loops=3)
                     Buckets: 4194304 (originally 2097152)  Batches: 1 (originally 1)  Memory Usage: 268736kB
                     ->  Parallel Bitmap Heap Scan on repo  (cost=26246.27..792112.25 rows=564262 width=46) (actual time=1081.285..2616.034 rows=937633 loops=3)
                           Recheck Cond: (((stars >= 5) AND (deleted_at IS NULL) AND (blocked IS NULL) AND (NOT archived) AND (NOT COALESCE(fork, false))) OR ((lower((name)::text) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)'::text) AND (deleted_at IS NULL) AND (blocked IS NULL)))
                           Rows Removed by Index Recheck: 2392
                           Heap Blocks: exact=197877
                           ->  BitmapOr  (cost=26246.27..26246.27 rows=1379405 width=0) (actual time=807.114..807.116 rows=0 loops=1)
                                 ->  Bitmap Index Scan on repo_stars_indexable  (cost=0.00..24377.83 rows=1246250 width=0) (actual time=530.668..530.669 rows=2661778 loops=1)
                                       Index Cond: (stars >= 5)
                                 ->  Bitmap Index Scan on erik_test_remove7  (cost=0.00..1191.33 rows=133155 width=0) (actual time=276.443..276.443 rows=170721 loops=1)
                                       Index Cond: (lower((name)::text) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)'::text)
 Planning Time: 1.272 ms
 Execution Time: 8602.761 ms
(25 rows)
```

**After**:

The following new queries (`not_cloned`/`cloned`, respectively) have the following query plans:

```sql
SELECT
    repo.id,
    repo.name,
    repo.stars
FROM repo
JOIN gitserver_repos gr ON gr.repo_id = repo.id AND gr.clone_status = 'not_cloned'
WHERE
   deleted_at IS NULL AND blocked IS NULL AND (
        (repo.stars >= 5 AND NOT COALESCE(fork, false) AND NOT archived)
        OR
        (lower(repo.name) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)')
    )
ORDER BY
    stars DESC NULLS LAST;
```

```
                                                                                                 QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Merge  (cost=257160.28..299573.44 rows=363516 width=46) (actual time=1496.149..1568.595 rows=9110 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Sort  (cost=256160.25..256614.65 rows=181758 width=46) (actual time=1486.820..1489.415 rows=3037 loops=3)
         Sort Key: repo.stars DESC NULLS LAST
         Sort Method: quicksort  Memory: 369kB
         Worker 0:  Sort Method: quicksort  Memory: 361kB
         Worker 1:  Sort Method: quicksort  Memory: 362kB
         ->  Parallel Hash Join  (cost=43756.92..240282.18 rows=181758 width=46) (actual time=741.508..1487.587 rows=3037 loops=3)
               Hash Cond: (gr.repo_id = repo.id)
               ->  Parallel Index Only Scan using gitserver_repos_not_cloned_status_idx on gitserver_repos gr  (cost=0.43..193899.41 rows=1000490 width=4) (actual time=0.048..436.436 rows=793001 loops=3)
                     Heap Fetches: 1063511
               ->  Parallel Hash  (cost=36702.84..36702.84 rows=564292 width=46) (actual time=734.558..734.559 rows=937636 loops=3)
                     Buckets: 4194304 (originally 2097152)  Batches: 1 (originally 1)  Memory Usage: 268704kB
                     ->  Parallel Index Only Scan using repo_dotcom_indexable_repos_idx on repo  (cost=0.43..36702.84 rows=564292 width=46) (actual time=0.050..243.893 rows=937636 loops=3)
                           Heap Fetches: 47941
 Planning Time: 6.194 ms
 Execution Time: 1569.188 ms
(18 rows)

Time: 1809.659 ms (00:01.810)
```

```sql
SELECT
    repo.id,
    repo.name,
    repo.stars
FROM repo
LEFT JOIN gitserver_repos gr ON gr.repo_id = repo.id AND gr.clone_status <> 'cloned'
WHERE
   deleted_at IS NULL AND blocked IS NULL AND (
        (repo.stars >= 5 AND NOT COALESCE(fork, false) AND NOT archived)
        OR
        (lower(repo.name) ~ '^(src\.fedoraproject\.org|maven|npm|jdk)')
    )
    AND gr.clone_status IS NULL
ORDER BY
    stars DESC NULLS LAST;
```

```
                                                                                                      QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Merge  (cost=259640.78..348900.51 rows=765030 width=46) (actual time=1575.369..2267.585 rows=2803738 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Sort  (cost=258640.75..259597.04 rows=382515 width=46) (actual time=1567.467..1634.918 rows=934579 loops=3)
         Sort Key: repo.stars DESC NULLS LAST
         Sort Method: quicksort  Memory: 113230kB
         Worker 0:  Sort Method: quicksort  Memory: 114762kB
         Worker 1:  Sort Method: quicksort  Memory: 113384kB
         ->  Parallel Hash Anti Join  (cost=180958.42..223171.75 rows=382515 width=46) (actual time=699.718..1316.673 rows=934579 loops=3)
               Hash Cond: (repo.id = gr.repo_id)
               ->  Parallel Index Only Scan using repo_dotcom_indexable_repos_idx on repo  (cost=0.43..36702.84 rows=564292 width=46) (actual time=0.042..186.521 rows=937636 loops=3)
                     Heap Fetches: 47947
               ->  Parallel Hash  (cost=168450.57..168450.57 rows=1000594 width=4) (actual time=688.859..688.860 rows=793022 loops=3)
                     Buckets: 4194304  Batches: 1  Memory Usage: 125856kB
                     ->  Parallel Index Only Scan using gitserver_repos_not_explicitly_cloned_idx on gitserver_repos gr  (cost=0.43..168450.57 rows=1000594 width=4) (actual time=0.046..412.618 rows=793022 loops=3)
                           Heap Fetches: 1064464
 Planning Time: 2.012 ms
 Execution Time: 2414.458 ms
(18 rows)

Time: 2433.936 ms (00:02.434)
```

## Test plan

Existing unit tests.